frontend/pyproject.toml
[project]
name = "drone-inspection-frontend"
version = "0.1.0"
description = "PyQt6 desktop frontend for the drone inspection backend."
requires-python = ">=3.12"
dependencies = [
    "PyQt6>=6.6",
    "requests>=2.31",
    "Pillow>=10.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8",
    "mypy>=1.13",
    "pytest>=8.3",
    "pytest-httpserver>=1.1",
    "types-requests>=2.31",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E", "F", "W",      # pycodestyle / pyflakes
    "I",                # isort
    "UP",               # pyupgrade
    "B",                # bugbear
    "SIM",              # simplify
    "RUF",              # ruff-specific
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B"]

[tool.mypy]
python_version = "3.12"
strict = true
files = ["app.py", "backend.py"]

[[tool.mypy.overrides]]
module = ["PyQt6.*", "PIL.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
